add medicamentos

+++ hooks/useUploadImg.ts
@@ -1,28 +1,31 @@
 import { PutBlobResult } from "@vercel/blob";
 
-export const uploadImage = async ({ file, url }: { file: File; url: string }): Promise<string> => {    
-    const deleteImg = await fetch(
-        `/api/images`,
-        {
-          method: 'DELETE',
-          body: JSON.stringify({
-            url,
-          }),
-        }
-      );
-
-    let imgUrl = '';
-    const response = await fetch(
-        `/api/upload?filename=${file.name}`,
-        {
-          method: 'POST',
-          body: file,
-        }
-      );
+export const uploadImage = async ({ file, url }: { file: File; url?: string }): Promise<string> => {    
   
-      const newBlob = await response.json();
-      const putBlobResult = newBlob as PutBlobResult;
-      imgUrl = putBlobResult.url;
+  if(url){
+    const deleteImg = await fetch(
+          `/api/images`,
+          {
+            method: 'DELETE',
+            body: JSON.stringify({
+              url,
+            }),
+          }
+        );
+  }
 
-    return imgUrl;
+  let imgUrl = '';
+  const response = await fetch(
+      `/api/upload?filename=${file.name}`,
+    {
+      method: 'POST',
+      body: file,
+    }
+  );
+  
+  const newBlob = await response.json();
+  const putBlobResult = newBlob as PutBlobResult;
+  imgUrl = putBlobResult.url;
+
+  return imgUrl;
 }
